+++ awc/src/ws.rs
@@ -382,7 +382,7 @@ impl WebsocketsRequest {
         if let Some(hdr_key) = head.headers.get(&header::SEC_WEBSOCKET_ACCEPT) {
             let encoded = ws::hash_key(key.as_ref());
 
-            if hdr_key.as_bytes() != &encoded {
+            if hdr_key.as_bytes() != encoded {
                 log::trace!(
                     "Invalid challenge response: expected: {:?} received: {:?}",
                     &encoded,