+++ tests/test_httpserver.rs
@@ -72,7 +72,7 @@ async fn test_start() {
 }
 
 #[cfg(feature = "openssl")]
-fn ssl_acceptor() -> std::io::Result<SslAcceptorBuilder> {
+fn ssl_acceptor() -> SslAcceptorBuilder {
     use openssl::{
         pkey::PKey,
         ssl::{SslAcceptor, SslMethod},
@@ -89,7 +89,7 @@ fn ssl_acceptor() -> std::io::Result<SslAcceptorBuilder> {
     builder.set_certificate(&cert).unwrap();
     builder.set_private_key(&key).unwrap();
 
-    Ok(builder)
+    builder
 }
 
 #[actix_rt::test]
@@ -102,7 +102,7 @@ async fn test_start_ssl() {
 
     thread::spawn(move || {
         let sys = actix_rt::System::new();
-        let builder = ssl_acceptor().unwrap();
+        let builder = ssl_acceptor();
 
         let srv = HttpServer::new(|| {
             App::new().service(web::resource("/").route(web::to(|req: HttpRequest| {