+++ awc/CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased - 2021-xx-xx
 ### Added
 * `ClientResponse::timeout` for set the timeout of collecting response body. [#1931]
+* `ClientBuilder::local_address` for bind to a local ip address for this client. [#2024]
 
 ### Changed
 * Feature `cookies` is now optional and enabled by default. [#1981]
@@ -15,6 +16,7 @@
 [#1931]: https://github.com/actix/actix-web/pull/1931
 [#1981]: https://github.com/actix/actix-web/pull/1981
 [#2008]: https://github.com/actix/actix-web/pull/2008
+[#2024]: https://github.com/actix/actix-web/pull/2024
 
 ## 3.0.0-beta.2 - 2021-02-10
 ### Added